Refactor avatar-label
