+++ vue/components/index.js
@@ -4,7 +4,6 @@ import { GlobalPlatforme } from "./parts/global-platforme/global-platforme.vue";
 import { OAuthPlatforme } from "./parts/oauth-platforme/oauth-platforme.vue";
 import { AlertPlatforme } from "./ui/alert-platforme/alert-platforme.vue";
 import { AvatarPlatforme } from "./ui/avatar-platforme/avatar-platforme.vue";
-import { AvatarLabelPlatforme } from "./ui/avatar-label-platforme/avatar-label-platforme.vue";
 import { ButtonColorPlatforme } from "./ui/button-color-platforme/button-color-platforme.vue";
 import { ButtonIconPlatforme } from "./ui/button-icon-platforme/button-icon-platforme.vue";
 import { ButtonPlatforme } from "./ui/button-platforme/button-platforme.vue";
@@ -44,7 +43,6 @@ const install = Vue => {
     Vue.component("oauth-platforme", OAuthPlatforme);
     Vue.component("alert-platforme", AlertPlatforme);
     Vue.component("avatar-platforme", AvatarPlatforme);
-    Vue.component("avatar-label-platforme", AvatarLabelPlatforme);
     Vue.component("button-color-platforme", ButtonColorPlatforme);
     Vue.component("button-icon-platforme", ButtonIconPlatforme);
     Vue.component("button-platforme", ButtonPlatforme);
@@ -83,7 +81,6 @@ export {
     OAuthPlatforme,
     AlertPlatforme,
     AvatarPlatforme,
-    AvatarLabelPlatforme,
     ButtonColorPlatforme,
     ButtonIconPlatforme,
     ButtonPlatforme,
