+++ vue/components/ui/avatar-platforme/avatar-platforme.vue
@@ -1,7 +1,27 @@
 <template>
-    <div class="avatar" v-bind:class="[{ notify: notify }, size]">
-        <img class="image" v-bind:src="imgUrl" alt="avatar" />
-        <div class="dot" v-bind:style="{ background: String(notifyColor) }" />
+    <div class="avatar" v-bind:class="[size]">
+        <div class="image-container" v-bind:class="[{ notify: notify }, size]">
+            <img class="image" v-bind:src="imgUrl" alt="avatar" />
+            <slot name="notify-dot">
+                <div
+                    class="dot"
+                    v-bind:style="{ 'border-color': notifyBorderColor, background: notifyColor }"
+                />
+            </slot>
+        </div>
+        <div class="text left" v-bind:class="[size]" v-if="title || subtitle">
+            <label-platforme class="title" v-if="title">
+                {{ title }}
+            </label-platforme>
+            <label-platforme class="subtitle" v-if="subtitle">
+                {{ subtitle }}
+            </label-platforme>
+        </div>
+        <div class="text right" v-bind:class="[size]">
+            <label-platforme class="content" v-if="content">
+                {{ content }}
+            </label-platforme>
+        </div>
     </div>
 </template>
 
@@ -10,45 +30,50 @@
 
 .avatar {
     display: inline-block;
-    position: relative;
     font-size: 0px;
 }
 
-.avatar .image {
+.image-container {
+    display: inline-block;
+    font-size: 0px;
+    position: relative;
+}
+
+.image-container .image {
     border-radius: 50%;
     height: auto;
 }
 
-.avatar.giant .image {
+.image-container.giant .image {
     height: 128px;
     width: 128px;
 }
 
-.avatar.giant .dot {
+.image-container.giant .dot {
     height: 17px;
     right: 9px;
     top: 8px;
     width: 17px;
 }
 
-.avatar.large .image {
+.image-container.large .image {
     height: 96px;
     width: 96px;
 }
 
-.avatar.large .dot {
+.image-container.large .dot {
     height: 12px;
     right: 6px;
     top: 7px;
     width: 12px;
 }
 
-.avatar.medium .image {
+.image-container.medium .image {
     height: 40px;
     width: 40px;
 }
 
-.avatar.medium .dot {
+.image-container.medium .dot {
     border-width: 0.9px;
     height: 6px;
     right: 3px;
@@ -56,12 +81,12 @@
     width: 6px;
 }
 
-.avatar.small .image {
+.image-container.small .image {
     height: 32px;
     width: 32px;
 }
 
-.avatar.small .dot {
+.image-container.small .dot {
     border-width: 0.9px;
     height: 4px;
     right: 3px;
@@ -69,12 +94,12 @@
     width: 4px;
 }
 
-.avatar.tiny .image {
+.image-container.tiny .image {
     height: 24px;
     width: 24px;
 }
 
-.avatar.tiny .dot {
+.image-container.tiny .dot {
     border-width: 0.3px;
     height: 3px;
     right: 2px;
@@ -82,12 +107,12 @@
     width: 3px;
 }
 
-.avatar.micro .image {
+.image-container.micro .image {
     height: 16px;
     width: 16px;
 }
 
-.avatar.micro .dot {
+.image-container.micro .dot {
     border-width: 0.1px;
     height: 2px;
     right: 1px;
@@ -95,7 +120,7 @@
     width: 2px;
 }
 
-.avatar .dot {
+.image-container .dot {
     background: #4b8dd7;
     border: 1px solid #ffffff;
     border-radius: 50%;
@@ -106,7 +131,64 @@
     width: 10%;
 }
 
-.avatar:not(.notify) .dot {
+.image-container:not(.notify) .dot {
+    display: none;
+}
+
+.text {
+    display: inline-flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    vertical-align: top;
+    flex-direction: column;
+    padding: 0px 8px 0px 8px;
+}
+
+.text.giant {
+    height: 128px;
+}
+
+.text.large {
+    height: 96px;
+}
+
+.text.medium {
+    height: 40px;
+}
+
+.text.small {
+    height: 32px;
+}
+
+.text.tiny {
+    height: 24px;
+}
+
+.text.micro {
+    height: 16px;
+}
+
+.text.left .title ::v-deep {
+    color: $dark;
+    font-size: 14px;
+    font-weight: 700;
+}
+
+.avatar.giant .text.left .title ::v-deep {
+    font-size: 80px;
+}
+
+.text.left .subtitle ::v-deep {
+    color: $dark;
+    font-weight: 500;
+    letter-spacing: 0.3px;
+}
+
+.text.right {
+    padding: 0px 0px 0px 0px;
+}
+
+.text:empty {
     display: none;
 }
 </style>
@@ -119,15 +201,31 @@ export const AvatarPlatforme = {
             type: String,
             default: "small"
         },
+        title: {
+            type: String,
+            default: null
+        },
+        subtitle: {
+            type: String,
+            default: null
+        },
+        content: {
+            type: String,
+            default: null
+        },
         imgUrl: {
             type: String,
-            default: ""
+            default: null
         },
         notify: {
             type: Boolean,
             default: false
         },
         notifyColor: {
+            type: String,
+            default: null
+        },
+        notifyBorderColor: {
             type: String,
             default: null
         }
